output/services/event-monitor/pyproject.toml
[project]
name = "venturestrat-event-monitor"
version = "1.0.0"
description = "Kafka event observer, audit trail, trace explorer, and topology API"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "VentureStrat Team"}]
dependencies = [
  "fastapi>=0.109.0",
  "uvicorn[standard]>=0.27.0",
  "pydantic>=2.5.0",
  "pydantic-settings>=2.1.0",
  "structlog>=24.1.0",
  "asyncpg>=0.29.0",
  "aiokafka>=0.10.0",
  "pyyaml>=6.0.0",
  "prometheus-client>=0.19.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=4.1.0",
  "httpx>=0.26.0",
  "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/event_monitor"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
